Changed data type in order to be able to remove cast.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@20081 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/kbd.c
@@ -57,7 +57,7 @@ BYTE _kbd_extended_key_tab[256] = {
 
 int kbd_handle_keydown(DWORD virtual_key, DWORD key_data)
 {
-    int kcode = (key_data >> 16) & 0xff;
+    long kcode = (key_data >> 16) & 0xff;
     int repeated = key_data & 0x40000000 ? 1 : 0;
 
     if (virtual_key == 0xff)
@@ -69,7 +69,7 @@ int kbd_handle_keydown(DWORD virtual_key, DWORD key_data)
     }
 
     if ( ! repeated ) {
-        keyboard_key_pressed((signed long)kcode);
+        keyboard_key_pressed(kcode);
     }
 
     return 0;
@@ -77,7 +77,7 @@ int kbd_handle_keydown(DWORD virtual_key, DWORD key_data)
 
 int kbd_handle_keyup(DWORD virtual_key, DWORD key_data)
 {
-    int kcode = (key_data >> 16) & 0xff;
+    long kcode = (key_data >> 16) & 0xff;
     int repeated = key_data & 0x40000000 ? 0 : 1;
 
     if (virtual_key == 0xff)
@@ -89,7 +89,7 @@ int kbd_handle_keyup(DWORD virtual_key, DWORD key_data)
     }
 
     if ( ! repeated ) {
-        keyboard_key_released((signed long)kcode);
+        keyboard_key_released(kcode);
     }
 
     return 0;
